Packaging for pdlua support files (WIP).
For now, these are located in Extra/pdlua. The external will look for
pd.lua in Library/Extra/pdlua in the plugdata library folder. This file
is needed for the external to work.

FIXME: It seems that at present, at first launch plugdata only populates
the plugdata library directory *after* the plugin has already been
instantiated and the setup routines of externals are called. So at first
launch pdlua will not be properly initialized. This can be worked
around by closing the first instance and creating another one.

The help files and examples all go into Documentation/13.pdlua. The
examples and the tutorial are in corresponding subfolders, the tutorial
examples can be found in tutorial/examples. Note that the help materials
also include a bunch of text files and Lua scripts, as well as the html
and pdf files of the tutorial. These files can't be opened directly in
plugdata at present, but they're still included so that you can open
them in your file manager.