Add pd-lua submodule, so that the external can be compiled in-tree.
Also adjusted the path to the external source in the cmake config and
scripts. In addition, plugdata now passes the entire Library/Extra path
to the setup routine, so that the external only needs to know that it
should look for support files in the pdlua subdir.

This will make it easier to move the support data around if the need
arises. The only files that need to be adjusted in such a case are
PdInstance.cpp and package_resources.py. The external itself will be
unaffected, as long as the pdlua subdir keeps its name.