Fixed botched logic in MidiDeviceManager::setMidiDeviceEnabled().
First, we need to check for an *input* device in the 2nd `else if`
branch, otherwise we're checking the status of the wrong device, which
causes changes to be ignored when toggling the state of an input device.

Also, we must *not* fall through to the `else` branch if the state check
of the the 2nd `else if` branch is false, because then we'd again be
checking the wrong device under the assumption that we're looking at an
*output* device now.

Finally, in the `else` branch we must *first* check for `shouldBeEnabled
!= isMidiDeviceEnabled(false, identifier)` to see whether anything needs
to be done at all, and *then* check just `shouldBeEnabled` to determine
whether the output device needs to be added or removed.

The result of all these glitches was that the input and output ports
would seemingly get enabled and disabled at random, and would only
accidentally be set to their proper values if you got (very) lucky. This
should all be fixed now.